% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInFiles.R
\name{FIF2dataframe}
\alias{FIF2dataframe}
\title{Output of `findInFiles` as a dataframe}
\usage{
FIF2dataframe(fif)
}
\arguments{
\item{fif}{the output of \code{\link{findInFiles}} used with the
option \code{output = "viewer+dataframe"}}
}
\value{
The results of \code{\link{findInFiles}} in a dataframe.
}
\description{
Returns the results of \code{\link{findInFiles}} in a dataframe, when the
  option \code{output = "viewer+dataframe"} is used.
}
\examples{
folder <- system.file("example", package = "findInFiles")
fif <- findInFiles("R", "function", root = folder, output = "viewer+dataframe")
FIF2dataframe(fif)
fif
}
