% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInFiles.R
\name{findInFiles-shiny}
\alias{findInFiles-shiny}
\alias{FIFOutput}
\alias{renderFIF}
\title{Shiny bindings for \code{findInFiles}}
\usage{
FIFOutput(outputId, width = "100\%", height = "400px")

renderFIF(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates a '\code{\link{findInFiles}}' widget}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{logical, whether \code{expr} is a quoted expression (with
\code{quote()})}
}
\description{
Output and render functions for using \code{findInFiles} within Shiny
applications and interactive Rmd documents.
}
