% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInFiles.R
\name{findInFiles}
\alias{findInFiles}
\title{<Add Title>}
\usage{
findInFiles(
  ext,
  pattern,
  depth = NULL,
  wholeword = FALSE,
  ignoreCase = FALSE,
  perl = FALSE,
  directory = "."
)
}
\description{
<Add Description>
}
