% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyFIF}
\alias{shinyFIF}
\title{Shiny application 'Find in files'}
\usage{
shinyFIF()
}
\value{
No returned value, just launches the Shiny application.
}
\description{
Launches a Shiny application allowing to run
  \code{\link{findInFiles}} and to navigate in the results.
}
\note{
The packages listed in the \strong{Suggests} field of the package
  description are required.
}
